Fix #3
Use instance variable instead of class variable for SKOS parsed elements.

We were setting parsed element as class variable using :attr_reader (class.send) on the SKOSInstance. This was setting the member for the whole class making the variable already defined in a next parse() call. A test was failing because of the presence of the member already defined. Due to the test failing, the member value was never set.